move some implementation of MaxEdgeLengthProcess to a source file
